% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd_add_phase_info.R
\name{dwd_add_phase_info}
\alias{dwd_add_phase_info}
\title{Add phase info to data file}
\usage{
dwd_add_phase_info(dwd_data)
}
\arguments{
\item{dwd_data}{a tibble containing DWD observation data to which station information shall be added = return of \code{\link{dwd_read}}.}
}
\value{
A tidyverse tibble containing the merged observation and phase data.
}
\description{
This function allows you to add phenological phase information to an observation data file. The matching DWD meta file containing relevant phase information is processed and joined.
}
\examples{
# read DWD file into R
my_file <- dwd_read(my_dir) # result: tidyverse tibble

# add phase information
my_file <- dwd_add_phase_info(my_file)

}
