% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd_clean.R
\name{dwd_clean}
\alias{dwd_clean}
\title{Clean relevant observation data}
\usage{
dwd_clean(dwd_data)
}
\arguments{
\item{dwd_data}{tibble containing DWD observation data to which station and phase information has been added already = return of functions dwd_add_phase/station_info(dir) or dwd_join_files(dwd_data_list).}
}
\value{
A tidyverse tibble containing the cleaned observation data.
}
\description{
This function allows you to clean observation data which is already joined with station and phase information by filtering relevant information. Duplicates are removed and irrelevant columns as well as incomplete or low-quality (based on quality flags) observations are deleted.
}
\examples{
# test example


}
