% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd_crop_list.R
\name{dwd_crop_list}
\alias{dwd_crop_list}
\title{List available DWD crops and abbreviations}
\usage{
dwd_crop_list()
}
\value{
A tidyverse tibble containing information on all available crops, their overall type and abbreviation used for other functions.
}
\description{
This function allows you to get and view a list of all crops with available phenological observations by the DWD. Names, overall crop type as well as abbreviation used in other functions are shown.
}
\examples{
# test example


}
