% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd_join_files.R
\name{dwd_join_files}
\alias{dwd_join_files}
\title{Add phase info to data file}
\usage{
dwd_join_files(dwd_data_list)
}
\arguments{
\item{dwd_data_list}{a list containing all observation data tibbles which shall be joined (returns of \code{\link{dwd_read}} or \code{\link{dwd_process}}).}
}
\value{
A tidyverse tibble containing all joined observation data.
}
\description{
This function allows you to join different tibbles containing observation data (usually immediate/annual reporters and/or recent/historic observations). Duplicates are removed.
}
\examples{
file_list <- c() # initialize list to wtore all processed files

# iterate through all files
for(file in dwd_data_list){

# add processing steps for single files here

# add processed file to list of all processed files
file_list <- c(file_list,my_file)

}

# join files
file_joined <- dwd_join_files(file_list)

}
