% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd_plot_2_ts.R
\name{dwd_plot_2_ts}
\alias{dwd_plot_2_ts}
\title{Plot one time-series per DWD station for two crops.}
\usage{
dwd_plot_2_ts(dwd_data_1, dwd_data_2, stat_ids, title, lab_1, lab_2)
}
\arguments{
\item{dwd_data_1}{tibble 1 containing DWD observation data (pre-processed, with station + phase information) which shall be plotted.}

\item{dwd_data_2}{tibble 2 containing additional DWD observation data.}

\item{stat_ids}{station ID(s) of station(s) which shall be plotted (optional). If empty, all stations are plotted.}

\item{title}{plot title (optional).}

\item{lab_1}{legend label data set 1 (optional).}

\item{lab_2}{legend label data set 2 (optional).}
}
\value{
The resulting time-series ggplot.
}
\description{
This function allows you to plot one time-series per DWD station for two DWD data sets (e.g. two different crops). Specific stations can be filtered beforehand. Furthermore, LOESS smoothening (local polynomial regression fitting) is applied.
}
