% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd_process.R
\name{dwd_process}
\alias{dwd_process}
\title{Process DWD phenology files}
\usage{
dwd_process(dir)
}
\arguments{
\item{dir}{directory of the folder which contains all DWD files to be processed (not ending with "/").}
}
\value{
A tidyverse tibble containing the processed and joined DWD phenology data.
}
\description{
This function allows you to process and join all relevant files of one crop at once. \cr
The following processing steps are applied to each file in the passed folder: \cr
\enumerate{
  \item read downloaded file into R (\code{\link{dwd_read}})
  \item add phase information from matching meta files (\code{\link{dwd_add_phase_info}})
  \item add station information from matching meta files (\code{\link{dwd_add_station_info}})
  \item clean data (\code{\link{dwd_clean}})
}
Then, the files are joined (\code{\link{dwd_join_files}}) and returned as one single tidyverse tibble.\cr
Note: The single functions for these steps can be accessed individually as well.
}
\examples{
## set directory where data files shall be saved
my_dir <- "C:/Users/.../my_folder"

## check available crops  and their abbreviations
dwd_crop_list()

## download
# both data + meta files
dwd_download("RBU",1900,2019,"JMSM",my_dir)

## create directory to folder containing files to be processed
folder_dir <- paste0(my_dir, "/RBU") # modify abbreviation accordingly

## process and join all files in folder
rbu_data <- dwd_process(dir)

}
