% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd_read.R
\name{dwd_read}
\alias{dwd_read}
\title{Read DWD phenology file into R}
\usage{
dwd_read(dir)
}
\arguments{
\item{dir}{directory where the file that shall be imported is located, including full file name.}
}
\value{
A tidyverse tibble containing the DWD phenology data.
}
\description{
This function allows you to import a DWD phenology file into R as a tidyverse tibble.
}
