% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd_stations_plot.R
\name{dwd_stations_plot}
\alias{dwd_stations_plot}
\title{Plot DWD stations on Germany map}
\usage{
dwd_stations_plot(dwd_data)
}
\arguments{
\item{dwd_data}{a tibble containing DWD observation data with station information.}
}
\description{
This function allows you to plot the location of all stations of a dwd data tibble on a map of Germany.
}
\examples{
# test example


}
