% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd_stations_shp.R
\name{dwd_stations_shp}
\alias{dwd_stations_shp}
\title{Save DWD stations as shape file}
\usage{
dwd_stations_shp(dwd_data, out_dir)
}
\arguments{
\item{dwd_data}{a tibble containing DWD observation data with station information.}

\item{out_dir}{directory of folder where created shape file shall be saved.}
}
\description{
This function allows you to save the location of all stations of a dwd data tibble as shape file (with +proj=longlat, +datum=WGS84).
}
\examples{
# test example


}
