% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ras_calc_mean.R
\name{ras_calc_mean}
\alias{ras_calc_mean}
\title{Calculate mean of each raster layer}
\usage{
ras_calc_mean(ras_stack, date_flag)
}
\arguments{
\item{ras_stack}{Raster stack to analyse.}

\item{date_flag}{A vector containing the date flags of each layer, e.g. observation years, which is appended to result. Ascending numbers are assigned if empty.}
}
\value{
A tibble with calculated mean pixel values per whole layer and according date flags.
}
\description{
This function allows you to calculate the mean of all pixel values per layer in a raster stack. A tibble with mean values and a column with a provided date flag is returned and can be used for forther analysis and plotting.
}
\examples{
# load processed data
ndvi_m_px_1 <- stack(".../phenoTS/example_scripts/MODIS_data_processed/ndvi_m_px_1.tif")

# mean of all pixels in layer
ndvi_m_1 <- ras_calc_mean(ndvi_m_px_1,date_flag=c(2000:2018))

}
