% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ras_calc_mean_points.R
\name{ras_calc_mean_points}
\alias{ras_calc_mean_points}
\title{Calculate mean of each raster layer for buffered points}
\usage{
ras_calc_mean_points(ras_stack, dir_shp, buffer, date_flag)
}
\arguments{
\item{ras_stack}{Raster stack to analyze.}

\item{dir_shp}{Full path to shape-file containing POIs (with file name and .shp) - e.g. created using dwd_stations_shp(). Provide unique point-IDs, column name = 'stat_id'!}

\item{date_flag}{A vector containing the date flags of each layer, e.g. observation years, which is appended to result. Ascending numbers are assigned if empty.}

\item{buffer_size}{Size of buffer around each point within which pixels are extracted and averaged (in meters).}
}
\value{
A tibble with calculated mean values per station and layer plus other information and date flags.
}
\description{
This function allows you to calculate the mean of all pixels within a buffer-zone around points of interest (POIs) for each layer in a raster stack. A tibble with mean values and additional columns with the provided date flags and other information is returned and can be used for forther analysis and plotting.
}
\examples{
# test example

}
