% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ras_plot_2_ts.R
\name{ras_plot_2_ts}
\alias{ras_plot_2_ts}
\title{Plot one time-series per scene or point for two data sets}
\usage{
ras_plot_2_ts(ras_data_1, ras_data_2, title, x_lab, y_lab, lab_1, lab_2)
}
\arguments{
\item{ras_data_1}{tibble 1 containing mean raster values which shall be plotted, e.g. result of ras_calc_mean() or ras_calc_mean_point().}

\item{ras_data_2}{tibble 2 containing mean raster values which shall be plotted.}

\item{title}{plot title (optional).}

\item{x_lab}{x-axis label (optional).}

\item{y_lab}{y-axis label (optional).}

\item{lab_1}{legend label data set 1 (optional).}

\item{lab_2}{legend label data set 2 (optional).}
}
\value{
The resulting time-series ggplot.
}
\description{
This function allows you to plot one time-series for two data sets, each per scene or extracted points - e.g. the results of ras_calc_mean() or ras_calc_mean_point(). Furthermore, LOESS smoothening (local polynomial regression fitting) is applied.
}
\examples{
# test example


}
