% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ras_plot_ts.R
\name{ras_plot_ts}
\alias{ras_plot_ts}
\title{Plot one time-series per scene or point}
\usage{
ras_plot_ts(ras_data, title, x_lab, y_lab)
}
\arguments{
\item{ras_data}{a tibble containing mean raster values which shall be plotted, e.g. result of ras_calc_mean() or ras_calc_mean_point().}

\item{title}{plot title (optional).}

\item{x_lab}{x-axis label (optional).}

\item{y_lab}{y-axis label (optional).}
}
\value{
The resulting time-series ggplot.
}
\description{
This function allows you to plot one time-series per scene or extracted points, e.g. the result of ras_calc_mean() or ras_calc_mean_point(). Furthermore, LOESS smoothening (local polynomial regression fitting) is applied.
}
\examples{
# test example


}
