% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd_filter.R
\name{dwd_filter}
\alias{dwd_filter}
\title{Filter TS-relevant data}
\usage{
dwd_filter(dwd_data, dwd_phase_id, obs_start, obs_end, obs_min)
}
\arguments{
\item{dwd_phase_id}{phase_id of selected phase according to DWD (not BBCH!) definition.}

\item{obs_start}{start of observation period (year).}

\item{obs_end}{end of observation period (year, included).}

\item{obs_min}{minimum number of available observations = years.}

\item{dwd_data_list}{a list containing all observation data tibbles which shall be joined (returns of dwd_read(dir) or dwd_add_phase/station_info(dir)).}
}
\value{
A tidyverse tibble containing all joined observation data.
}
\description{
This function allows you to filter observation data relevant to your time-series analysis. A single phase and observation period is selected. Observations of stations closed out of or within this period are deleted. All observations of a station with less than a selected number of observations is available are removed.
}
\examples{
# test example


}
