% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dwd_join_files.R
\name{dwd_join_files}
\alias{dwd_join_files}
\title{Add phase info to data file}
\usage{
dwd_join_files(dwd_data_list)
}
\arguments{
\item{dwd_data_list}{a list containing all observation data tibbles which shall be joined (returns of dwd_read(dir) or dwd_add_phase/station_info(dir)).}
}
\value{
A tidyverse tibble containing all joined observation data.
}
\description{
This function allows you to join different tibbles containing observation data (usually immediate/annual reporters and/or recent/historic observations). Duplicates are removed.
}
\examples{
# test example


}
