% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ras_calc_mean.R
\name{ras_calc_mean}
\alias{ras_calc_mean}
\title{Calculate mean of each raster layer}
\usage{
ras_calc_mean(ras_stack, date_flag)
}
\arguments{
\item{ras_stack}{Raster stack to analyse.}

\item{date_flag}{A vector containing the date flags of each layer, e.g. observation years, which is appended to result. Ascending numbers are assigned if empty.}
}
\value{
A tibble with calculated mean values per layer and according date flags.
}
\description{
This function allows you to calculate the mean of all pixels per layer in a raster stack. A tibble with mean values and a column with a provided date flag is returned and can be used for forther analysis and plotting.
}
\examples{
# test example

}
